Revert ":bug: ::  글 작성 기능 api 연결, imgurl 삽입 안됨"
