+++ components/Main/index.tsx
@@ -7,7 +7,7 @@ const Main = () => {
   return (
     <div className="">
       <div className="flex pt-[130px] pl-[200px]">
-        {/* <LostpostList /> */}
+        <LostpostList />
         <Profile />
       </div>
     </div>
